优化模块组织.3 声明式渲染器管理

+++ Camera.hpp
@@ -48,15 +48,15 @@ public:
     {
         if (firstMouse)
         {
-            lastX = xpos;
-            lastY = ypos;
+            lastX = static_cast<float>(xpos);
+            lastY = static_cast<float>(ypos);
             firstMouse = false;
         }
 
-        float xoffset = xpos - lastX;
-        float yoffset = lastY - ypos;
-        lastX = xpos;
-        lastY = ypos;
+        float xoffset = static_cast<float>(xpos - lastX);
+        float yoffset = static_cast<float>(lastY - ypos);
+        lastX = static_cast<float>(xpos);
+        lastY = static_cast<float>(ypos);
 
         xoffset *= sensitivity;
         yoffset *= sensitivity;
